+++ src/main/java/Moves/Empty.java
@@ -1,8 +1,12 @@
-package Server.Moves;
+package Moves;
 
 public class Empty implements Move {
     @Override
     public String toString() {
         return "EMPTY";
+    }
+
+    public String pretty(){
+        return "-";
     }
 }
