+++ src/main/java/Moves/GiveUp.java
@@ -1,8 +1,12 @@
-package Server.Moves;
+package Moves;
 
 public class GiveUp implements Move {
     @Override
     public String toString() {
         return "GIVEUP";
+    }
+
+    public String pretty(){
+        return "Poddał się";
     }
 }
