+++ src/main/java/Moves/Pass.java
@@ -1,8 +1,12 @@
-package Server.Moves;
+package Moves;
 
 public class Pass implements Move {
     @Override
     public String toString() {
         return "PASS";
+    }
+
+    public String pretty(){
+        return "Pass";
     }
 }
